pyproject.toml

[tool]
[tool.poetry]
name = "hav"
version = "0.1.0"
description = "The Himalaya Archive Vienna."
authors = [
    "Sean McAllister <sean@brickwall.at>"
]
packages = [
 { include = "hav", from = "src" },
]


[tool.poetry.dependencies]
python = ">=3.8,<4.0"
channels = "^3.0.4"
channels-redis = "*"
dj-database-url = "*"
django = "^3.2"
django-cors-headers = "*"
django-environ = "==0.4.4"
django-extensions = "==2.1.5"
django-model-utils = "^4.1.1"
django-redis = "*"
django-rq = "*"
django-treebeard = "^4.5.1"
django-webpack-loader = "==1.4.1"
djangorestframework = "^3.11"
graphene-django = "*"
ipython = "*"
pick = "*"
pillow = "*"
pip = "*"
psycopg2 = "*"
pycountry = "*"
pyexiftool = {git = "https://github.com/whav/pyexiftool.git"}
pylint = "*"
python-dotenv = "==0.7.1"
pyvips = "^2.1.15"
redis = "==3.0.1"
requests = "*"
sentry-sdk = "*"
uwsgi = "*"
whitenoise = "*"
meilisearch = "^0.16.2"
pydantic = "^1.7.3"
rawpy = "^0.17.0"
imageio = "^2.16.1"
hashids = "^1.3.1"
#Django = "^3.2.5"
markdown-it-py = {extras = ["plugins"], version = "^1.1.0"}
cryptography = "3.4.7"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "*"
ipdb = "*"
pytest = "*"
pytest-django = "*"
black = {version = "^20.8b1", allow-prereleases = true}
jedi = "0.17.2"
pre-commit = "^2.17.0"
